+++ api/control/telefone_control.go
@@ -100,7 +100,7 @@ func (server *Server) GetTelefoneByID(w http.ResponseWriter, r *http.Request) {
 	telefone := models.Telefone{}
 
 	//	telefoneGotten recebe o dado buscado no banco de dados
-	telefoneGotten, err := telefone.FindTelefoneByID(server.DB, uint64(telefoneID))
+	telefoneGotten, err := telefone.FindTelefoneByID(server.DB, telefoneID)
 
 	if err != nil {
 		responses.ERROR(w, http.StatusBadRequest, fmt.Errorf("[FATAL] It couldn't find by ID, %v\n", err))
@@ -165,7 +165,7 @@ func (server *Server) DeleteTelefone(w http.ResponseWriter, r *http.Request) {
 
 	/* 	Para o caso da funcao 'delete' apenas o erro nos eh necessario
 	Caso nao seja possivel deletar o dado especificado tratamos o erro*/
-	_, err = telefone.DeleteTelefone(server.DB, uint64(telefoneID))
+	_, err = telefone.DeleteTelefone(server.DB, telefoneID)
 	if err != nil {
 		formattedError := config.FormatError(err.Error())
 		responses.ERROR(w, http.StatusInternalServerError, fmt.Errorf("[FATAL] it couldn't delete in database , %v\n", formattedError))
